+++ src/entry/main.ts
@@ -1,9 +1,9 @@
 import encodedWasm from '../blackbox-log.wasm?inline';
-import workerUrl from '../worker?url';
 
 export * from './common';
 export * from '../async';
 export * from '../sync';
+export { default as worker } from '#worker?inline';
 
 export async function getWasm(): Promise<WebAssembly.Module> {
 	const decoded = atob(encodedWasm);
@@ -15,5 +15,3 @@ export async function getWasm(): Promise<WebAssembly.Module> {
 
 	return WebAssembly.compile(bytes);
 }
-
-export const worker = workerUrl;