+++ src/env.d.ts
@@ -1,0 +1,4 @@
+declare module '#worker?inline' {
+	const worker: string;
+	export default worker;
+}