+++ vite.config.js
@@ -1,7 +1,10 @@
 import { readFile } from 'node:fs/promises';
+import { resolve } from 'node:path';
 import { defineConfig } from 'vite';
 import { viteStaticCopy as staticCopy } from 'vite-plugin-static-copy';
 import { dependencies } from './package.json';
+
+const isDev = process.env.NODE_ENV === 'development';
 
 // Vite default + wasm sign-extension, multi-value, and bulk memory
 export const target = [
@@ -41,6 +44,15 @@ export default defineConfig({
 					code: `export default '${wasm}'`,
 					map: { mappings: '' },
 				};
+			},
+		},
+		{
+			name: 'inline-worker',
+			async resolveId(id) {
+				if (id === '#worker?inline') {
+					const file = isDev ? './src/worker.ts' : './dist/worker.js';
+					return await this.resolve(resolve(`${file}?url`));
+				}
 			},
 		},
 		staticCopy({